src/AIConnector/requirements.txt
colorlog==6.9.0
wcwidth==0.2.13
python-dotenv==1.0.1
websockets~=15.0
azure-messaging-webpubsubservice==1.2.1
